refactor(diff): move shared diff helpers
