+++ git-productivity-analyzer/src/cmd/commit_size/run.rs
@@ -5,5 +5,6 @@ use crate::Globals;
 
 pub async fn run(args: Args, globals: &Globals) -> Result<()> {
     let opts: Options = args.into();
+    opts.validate()?;
     run_with_analyzer(opts, globals, |a, summary| a.print_summary(summary)).await
 }
