+++ git-productivity-analyzer/src/sdk/churn/analyzer.rs
@@ -1,7 +1,7 @@
 use std::collections::BTreeMap;
 
-use super::diff_utils::commit_trees;
-use super::diff_utils::{compute_diff_lines, configure_changes, create_changes};
+use super::diff_utils::compute_diff_lines;
+use crate::sdk::diff::{commit_trees, configure_changes, create_changes};
 use crate::{error::Result, Globals};
 use bytecount::count;
 use gix::bstr::{BStr, ByteSlice};
