+++ git-productivity-analyzer/src/sdk/churn/diff_utils.rs
@@ -1,20 +1,5 @@
-use crate::error::Result;
-use gix::{
-    bstr::BStr,
-    object::tree::diff::{Change, Platform},
-    Tree,
-};
-use miette::Report;
-
-pub fn create_changes<'a, 'repo>(tree: &'a Tree<'repo>) -> Result<Platform<'a, 'repo>> {
-    tree.changes().map_err(|e| Report::msg(e.to_string()))
-}
-
-pub fn configure_changes(changes: &mut Platform<'_, '_>) {
-    changes.options(|opts| {
-        opts.track_filename().track_path().track_rewrites(None);
-    });
-}
+use gix::bstr::BStr;
+use gix::object::tree::diff::Change;
 
 pub fn compute_diff_lines(
     prev_mode: gix::object::tree::EntryMode,
@@ -39,21 +24,4 @@ pub fn compute_diff_lines(
     } else {
         None
     }
-}
-
-pub fn commit_trees<'repo>(
-    repo: &'repo gix::Repository,
-    commit_id: gix::ObjectId,
-    parent: Option<gix::ObjectId>,
-) -> (gix::Tree<'repo>, gix::Tree<'repo>) {
-    let to = repo
-        .find_object(commit_id)
-        .ok()
-        .and_then(|o| o.peel_to_tree().ok())
-        .unwrap_or_else(|| repo.empty_tree());
-    let from = parent
-        .and_then(|id| repo.find_object(id).ok())
-        .and_then(|c| c.peel_to_tree().ok())
-        .unwrap_or_else(|| repo.empty_tree());
-    (from, to)
 }
