+++ git-productivity-analyzer/src/sdk/churn/mod.rs
@@ -2,4 +2,3 @@ mod analyzer;
 mod diff_utils;
 
 pub use analyzer::Options;
-pub use diff_utils::{commit_trees, configure_changes, create_changes};
